requirements.txt
# --- Core Deep Learning ---
# TensorFlow enthält Keras bereits, eine separate Installation ist oft nicht nötig, 
# sorgt aber für Klarheit bei den Imports.
tensorflow>=2.15.0
keras>=2.15.0
numpy>=1.24.3

# --- Image Processing ---
# Pillow ist essenziell für dein load_data Skript
pillow>=10.0.0
opencv-python-headless>=4.8.0

# --- Data Science Tools ---
# Pandas für das Label-Handling, Scikit-Learn für den Train-Test-Split
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# --- Visualization (Optional für Fehleranalyse) ---
matplotlib>=3.8.0
seaborn>=0.13.0

# --- Web Interface (Falls du eine Demo bauen willst) ---
# Streamlit ist perfekt, um das Modell im Browser zu zeigen
streamlit>=1.30.0

# --- Utilities ---
tqdm>=4.66.0